requirements.txt
django==3.0.7
gunicorn==20.0.4
whitenoise==5.2.0
Pillow==7.1.2
dj-database-url==0.5.0
psycopg2-binary==2.8.6
mysqlclient==2.0.1
djangorestframework==3.12.4
django-rest-swagger==2.2.0